+++ src/CountrySelect.php
@@ -1,59 +1,34 @@
 <?php
-/**
- * Country Select plugin for Craft CMS 3.x
- *
- * Country select field type.
- *
- * @link      https://github.com/lukeyouell
- * @copyright Copyright (c) 2018 Luke Youell
- */
 
 namespace lukeyouell\countryselect;
 
-use lukeyouell\countryselect\fields\CountrySelectField as CountrySelectField;
-use lukeyouell\countryselect\fields\CountrySelectMultiField as CountrySelectMultiField;
-
 use Craft;
 use craft\base\Plugin;
-use craft\services\Plugins;
-use craft\events\PluginEvent;
 use craft\services\Fields;
 use craft\events\RegisterComponentTypesEvent;
 
+use lukeyouell\countryselect\fields\CountrySelectCheckboxes;
+use lukeyouell\countryselect\fields\CountrySelectDropdown;
+use lukeyouell\countryselect\fields\CountrySelectMultiSelect;
+use lukeyouell\countryselect\fields\CountrySelectRadioButtons;
+
 use yii\base\Event;
 
-/**
- * Class CountrySelect
- *
- * @author    Luke Youell
- * @package   CountrySelect
- * @since     1.0.0
- *
- */
 class CountrySelect extends Plugin
 {
     // Static Properties
     // =========================================================================
 
-    /**
-     * @var CountrySelect
-     */
     public static $plugin;
 
     // Public Properties
     // =========================================================================
 
-    /**
-     * @var string
-     */
     public $schemaVersion = '1.0.0';
 
     // Public Methods
     // =========================================================================
 
-    /**
-     * @inheritdoc
-     */
     public function init()
     {
         parent::init();
@@ -63,31 +38,11 @@ class CountrySelect extends Plugin
             Fields::class,
             Fields::EVENT_REGISTER_FIELD_TYPES,
             function (RegisterComponentTypesEvent $event) {
-                $event->types[] = CountrySelectField::class;
-                $event->types[] = CountrySelectMultiField::class;
+                $event->types[] = CountrySelectCheckboxes::class;
+                $event->types[] = CountrySelectDropdown::class;
+                $event->types[] = CountrySelectMultiSelect::class;
+                $event->types[] = CountrySelectRadioButtons::class;
             }
-        );
-
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function (PluginEvent $event) {
-                if ($event->plugin === $this) {
-                }
-            }
-        );
-
-        Craft::info(
-            Craft::t(
-                'country-select',
-                '{name} plugin loaded',
-                ['name' => $this->name]
-            ),
-            __METHOD__
         );
     }
-
-    // Protected Methods
-    // =========================================================================
-
 }